flask_app/requirements_original.txt
flask
gunicorn
webcolors==1.11.1
absl-py==0.10.0
astor==0.8.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
gast==0.2.2
google-auth==1.21.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.5.2
Markdown==3.2.2
MarkupSafe==1.1.1
numpy==1.19.1
oauthlib==3.1.0
opencv-python==4.1.2.30
opt-einsum==3.3.0
Pillow==7.2.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
six==1.15.0
termcolor==1.1.0
tqdm==4.48.2
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
